tidy up under google directory 1) add question description to each solution class 2) move them to relevent package 3) change some main to test 4) seperate test methods into test classes for part of them (the remaining will be done later)

+++ src/leetcode/success/bit/HammingDistance.java
@@ -4,7 +4,7 @@ package leetcode.success.bit;
 
 class HammingDistance {
 	public int hammingDistance(int x, int y) {
-		int xor=(int)(x^y);
+		int xor= x^y;
 		int dist=0;
 
 		do{
